authenticata a subroute

+++ models/authentication_response.go
@@ -3,6 +3,6 @@ package models
 type AuthenticationResponse struct {
 	User
 	Token     string `json:"token"`
-	ExpiresAt int    `json:"expires_at"`
+	ExpiresAt int64  `json:"expires_at"`
 	Message   string `json:"message"`
 }
